fish tank and rock paper scissors
FishTank: this code was written to experiment with jquery to create a large image in a webpage with interactive buttons hidden in the image that would take the user to another page. Buttons had to be able to change size with image on webpage. There was difficulty making the buttons appear when user hovered over it because there isn't css that allows shapes in image maps to change color or border... (to my knowledge anyway. I will research it more later if I come back to this)
The reason I experimented with image maps was because the museum I volunteer at asked for some website help and asked if this was possible.

rockPaperScissor: code that was written with another student to help with a homework assignment in their introductory computer science course in C++